[Chore] rename legacy auth api filename
